Add files via upload
To activate it must comment out SwordAttackController and uncomment ArrowAttackController in PlayerAttackingSpriteController class. Some hard coded numbers for offset and all code in PlayerArrowAttackingSpriteController Class... I will break into 4 instances based on facing direction and major refactor in the future.... Still buggy on the offset and arrow duration... but it shoots!